feat: insert next auth
